Fixed rendering of incoming quoted contacts.
Previously, when we found a match for a quoted message, we'd replace the
incoming quoted text with the body of the matched message. Most of the
time this is fine, but shared contacts don't have body text! So we were
erasing it. I just switched it to always use the text on the incoming
quote proto.